구조 검증 및 연동성 검증 1차 수정 완료

+++ spec/Schema_request.py
@@ -16,8 +16,8 @@ cmgvieyak001b6cd04cgaawmm_CameraProfiles_in_schema = {}
 # StoredVideoInfos
 cmgvieyak001b6cd04cgaawmm_StoredVideoInfos_in_schema = {
     "timePeriod": {
-    "startTime": float,
-    "endTime": float,
+    "startTime": int,
+    "endTime": int,
 },
     OptionalKey("camList"): [{
     "camID": str,
@@ -36,8 +36,8 @@ cmgvieyak001b6cd04cgaawmm_StreamURLs_in_schema = {
 cmgvieyak001b6cd04cgaawmm_ReplayURL_in_schema = {
     "camList": [{
     "camID": str,
-    "startTime": float,
-    "endTime": float,
+    "startTime": int,
+    "endTime": int,
     "streamProtocolType": str,
 }],
 }
@@ -47,14 +47,14 @@ cmgvieyak001b6cd04cgaawmm_RealtimeVideoEventInfos_in_schema = {
     "camList": [{
     "camID": str,
 }],
-    "duration": float,
+    "duration": int,
     "transProtocol": {
     "transProtocolType": str,
     "transProtocolDesc": str,
 },
     "eventFilter": str,
     "classFilter": str,
-    "startTime": float,
+    "startTime": int,
 }
 
 # RealtimeVideoEventInfos WebHook OUT Schema
@@ -66,13 +66,13 @@ cmgvieyak001b6cd04cgaawmm_RealtimeVideoEventInfos_webhook_out_schema = {
 # StoredVideoEventInfos
 cmgvieyak001b6cd04cgaawmm_StoredVideoEventInfos_in_schema = {
     "timePeriod": {
-    "startTime": float,
-    "endTime": float,
+    "startTime": int,
+    "endTime": int,
 },
     "camList": [{
     "camID": str,
 }],
-    "maxCount": float,
+    "maxCount": int,
     "classFilter": str,
     "eventFilter": str,
 }
